requirements.txt
asgiref==3.7.2
attrs==23.2.0
channels==4.0.0
click==8.1.7
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
gunicorn==21.2.0
h11==0.14.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.5.2
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.17.1
sqlparse==0.4.4
uritemplate==4.1.1
uvicorn==0.27.1
